feat: add config file

+++ internal/keyring/account.go
@@ -2,6 +2,6 @@ package keyring
 
 import "fmt"
 
-func GetYubikeyAccount(keyID uint64, name string) string {
+func GetYubikeyAccount(keyID uint32, name string) string {
 	return fmt.Sprintf("yubikey:%d:%s", keyID, name)
 }
